Create The Fox of Fallacy.html
Cosmic@scowl : "you're an agent Mr. Skechit - an eponymy only#..As i'm..the Jackal of Realisation,#,..Sazare un..long,Tracey shakes..lot,#,..but what do you.. coalesce is the query..to disprove.. that it's not between a blatant lie and a speculation on..evaluable autonomy..a counter or a lie in itself..and these scheme of lying Mr. Skechit that we've just bee tricked into how do you formulate self in this#..Phantom : The Meaning of Lie,#,.."
Tracey@thoughtful : "but a lie or a truth only begins with an assumption of exclusivity over me from what is observed.."
Cosmic@angry : "ai ai zip it#..As i'm..the Fox of Fallacy,#,..Tracey+Sazare un..long,#,.."